Update statuses of HTTPRoutes (#60)
This commit adds a new component - status.Updater - which updates
statuses of HTTPRoutes. In the future, it will update statuses of the
other supported resources.

The added component has a few limitations (described in the comments)
that are required to be addressed before the Gateway is considered to be
ready for production use.