Support processing multiple HTTP listeners
Problem:
The Gateway needs to process a Gateway resource. Currently, that is
already supported by the 'state' package. However, that support is
limited. For example, the package doesn't allow processing  multiple
listeners. More over, the approach used in the 'state' package is hard
to extend to support multiple listeners. 
This prompts development of a better solution.

As a first deliverable, we want to support processing multiple HTTP
listeners assuming they all bind to port 80. 
We also assume that the Gateway only supports a single Gateway resource
with known namespace and name.

Solution - replace the existing types of the state package with
the new ones:
- Introduce Configuration, which is an internal representation of the
Gateway configuration. 
- Introduce Statuses, which holds status-related information about
processed resources.
- Introduce ChangeProcessor, which processes changes to Gateway API
resources and returns new Configuration and Statuses.