Enforce that the SNI name matches the host name
Problem: Nginx does not enforce that the SNI name matches the host name for SSL
requests, which can lead to unexpected behavior and breaks the
recommendation of the gateway API spec.

Fix: in every SSL server block, we now check that the variables $ssl_server_name
and $host are equal. If they are not, we return a 421 misdirected request.