Generate a default SSL server for select HTTPS listener hostname (#173)
Generate a default SSL server for select listener hostnames

Problems:

(1) If a valid HTTPS listener is configured but has no attached routes, a
request to the listener's hostname will result in an SSL handshake error.

(2) If a valid HTTPS listener with a catch-all hostname is configured, a request
to any hostname that does not match an existing route will result in an SSL
handshake error.

After discussion, we decided that it doesn't make sense to reject the SSL
handshake when there's a valid HTTPS listener configured for the hostname. The
desired behavior is to return a 404.

Solution: Generate an SSL server block for each listener described above that
terminates the TLS connection and returns a 404.