Update handling of query params in njs
Updating NGINX to 1.23.1 image brought changes with njs-0.7.6, which now
supports examining all values of a query param.

The commit updates njs code for matching query params so that in the case
of multiple values of a query param, the first value is used. This is
consistent with the previous behavior and the Gateway API spec.

Additionally, the unit tests for 'paramsMatch' were fixed and updated
to cover multiple param values and different case of a param name.