Implement request redirect filter in HTTPRoute rule (#218)
This commit implements the request redirect filter as part of the routing
rule in the HTTPRoute.

A common use-case for a request redirect is redirecting HTTP requests
to HTTPS. The commit updates the HTTPS termination example to include
HTTPS redirect configuration.

Notes:
- The experimental 'path' field of 'requestRedirect' is out of scope.
- The validation of the fields of `requestRedirect` is not implemented.
It is left to be done in a separate component responsible for validation
with FIXMEs added to the relevant code locations.
- If multiple redirect filters are configured, NGINX Kubernetes Gateway
will choose the first one and ignore the rest.
- NGINX will always redirect a request even if the request has already
been redirected. Thus, any backendRefs defined in the routing rule will
be ignored. However, that "always redirect" behavior is not specified
by the Gateway API. As a result, we might need to change our
implementation if different behavior becomes specified by the Gateway API
in the future.