Refactor controllers and implementations
This commit refactors controllers and implementations. Previously, all
controllers and implementations looked almost the same. The commit
replaces all controllers with a single controller that handles all
the types. Similarly, the commit replaces all implementations with
a single one that handles all the types.

The new code relies on generics