Refactor reconcilers and implementations
Previously, all reconcilers and implementations looked almost the same.

This commit:
- Replaces all reconcilers with a single one.
- Removes all implementations. Now the reconciler sends upsert and
delete events to the event channel instead of the implementations.
The sending to the events part of the reconciler doesn't have
the shutdown bug (a possibility to hang while sending to the channel
during the shutdown) all implementations shared.
- Removes the sdk package.

Additionally, the Manager now uses the same logger as the other
components.