Support NoMatchingListenerHostname reason in HTTPRoute status (#290)
If an HTTPRoute references a listener with the hostname that does not
match any hostnames of the HTTPRoute, NGINX Kubernetes Gateway will
set the condition "Accepted" with the status "False" and the reason
"NoMatchingListenerHostname" in the status of the HTTPRoute.

The commit introduces a new type RouteCondition. It is used to capture
status-related conditions when building the graph. It will be used to
capture other conditions too, including when building configuration.

Note: the implementation required implementing a few additional
conditions. In one case, to wait until the corresponding reasons
(v1beta1.RouteConditionReason) are added to the Gateway API and in the
other case, to not increase the scope, those conditions are left to be
implemented (FIXMEs).