Improve performance of endpoints deduplication (#299)
We use a map as a set to deduplicate endpoints. Before deduplicating, 
we can calculate the total number of the endpoints in the input
and assume that most of those endpoints are unique. Then, we can use
that number when initializing the map. That will improve 
the performance, as it will help to reduce the cost of growing the map
to accommodate all the endpoints.

The benchmarks are included.