Use strict versions in package.json and align CI
With `^` it means that we can get any version up to the latest minor or patch version. For example "prettier": "^2.8.3" it means at least 2.8.3 but anything before to 3.0.0, with this change we get the exact same version all the time.
The CI now reads package.json to get the same version of prettier to run.