Remove redundant warnings
The warnings reported by dataplane.BuildConfiguration() are now
redundant - they are reported as Conditions in the statuses of
Gateway API resources, which includes warnings-related cases for:
* invalid listener in an HTTPRoute
* unresolved backend ref an HTTPRoute

This commit removes the code, which generates and reports warnings.

Additionally, the commit fixes a related test TestBuildConfiguration
case "invalid listener": if a listener is invalid, it cannot include
any Routes and AcceptedHostnames, so those fields are removed.

Fixes https://github.com/nginxinc/nginx-kubernetes-gateway/issues/467