Exact PathMatch support for HTTPRoutes (#603)
* Exact PathMatch support for HTTPRoutes

Add support for Exact PathMatch in an HTTPRoute. Internal locations now include the type of path (prefix, exact, regex) in the path name to distinguish between the possibility of the same path name being used with different path match types.

nginx conf is updated to include "= " in the location path if that path has been defined as "Exact".