Set Accepted condition type on Gateway status (#633)
* Set Accepted condition type on Gateway status

We always report the Accepted condition on the Gateway status. Its value depends on the validity of its spec and its Listeners. If the Gateway spec is invalid, we report the Accepted/False/Invalid Gateway condition but do not report Listener statuses. In all other cases, the Listener statuses will be reported. For routes that reference an invalid Gateway, we set the Route condition Accepted/False/InvalidGateway.

* Reduce the complexity of bindRouteToListeners

The gocyclo linter complained about the complexity of the bindRouteToListeners function. This commit refactors this
function to reduce its complexity.