Set Gateway Programmed condition
When the Gateway configuration has been successfully uploaded to nginx, we will set the status Programmed to true. If there's any error with updating nginx, then we set Programmed to false. We'll also set a negative status on the HTTPRoutes in the system to explain that the Gateway is not programmed.

Moved the buildStatuses logic into the status package and removed redundant tests.