Support cross-namespace BackendRefs in HTTPRoutes (#806)
Allow HTTPRoutes to reference Backends in different namespaces if a ReferenceGrant permits it.

Problem: NKG does not allow HTTPRoutes to reference Backends in different namespaces.

Solution: Allow HTTPRoutes to reference Backends in different namespaces if a ReferenceGrant 
permits it. Both SecretObjectReferences and BackendObjectReferences are resolved through the referenceGrantResolver. The referenceGrantResolver transforms the map of ReferenceGrants into
a structure that allows for O(n) lookup of references.