Set proxy host header to the exact value of the request host header (#827)
Problem: The Gateway API expects the response host header to match the 
exact value of the request host header. We set the proxy host header to the 
$host nginx variable, which does not include any characters after ":". 
This causes a conformance test to fail where the host header contains a port.

Solution: Set the proxy host header to $http_host nginx variable when it is 
non-empty. This variable contains the unmodified host header of the request. 
However, in an HTTP/1.0 request, it's possible that $http_host can be empty. 
In this case, we will use the value of $host. 
See http://nginx.org/en/docs/http/ngx_http_core_module.html#var_host.