Check for changes in generated files in GitHub workflow (#844)
Problem: If we forget to run make generate, the out-of-date code 
can be committed to the main branch. 

Solution: Add a step to the Checks and Variables job in the 
CI workflow that checks if the generated files have changed. 
If they have changed, the job will fail, and the PR will be blocked.