Dynamic control plane configuration support
Problem: We want to be able to change control plane settings dynamically, to avoid having to restart NKG and lose valuable state.

Solution: Introducing a new CRD that is initialized and created on startup, NginxGateway. This CRD can be updated by the user to dynamically change the state of the control plane. Right now we simply support changing the logging level. The controller will revert to using default values if the CRD is not detected.