Build and use custom nginx container (#934)
Replace the publicly available nginx image with a custom nginx image.

Problem: Using the publicly available nginx image requires users to create 
ConfigMaps for the nginx.conf file and the njs module and mount them to 
the NKG Pod as volumes. This pattern is not extensible and adds extra 
steps for developers and users. Additionally, an init container is required 
in order to set up the nginx config environment.

Solution: Build and use a custom nginx container. The nginx.conf and njs module 
are now baked into the nginx image. This eliminates the need for ConfigMaps. 
The config directories /etc/nginx/conf.d and /etc/nginx/secrets are created as 
volumes and mounted to the Pod with a group ID 1001. This allows the control 
plane to write to the directories and nginx to read from them. Both the nginx 
and nginx-gateway processes run under group ID 1001 but have different 
user IDs (101 and 102). The nginx container runs as user 101 instead of root
and runs with the minimum set of capabilities.