Fix CRD manifest installation
Problem: Deploying a CRD definition and an instantiation of that CRD in the same kubectl command will fail due to the kubectl cache not being updated for the existence of the CRD definition. They need to be run in separate commands.
Solution: Separate the CRDs into a different file and install prior to installing the rest of the gateway resources.