Add deployment manifests as release artifacts (#1092)
Problem: As a user, I want an easy way to install the NGF manifests. Currently, I have to clone the repository.

Solution: By including the manifests as part of a release, the repo no longer has to be cloned to install NGF. Note: the service definitions are not uploaded and therefore still require a clone if they are to be installed.