Add event processing histogram metric (#1134)
Problem: NGF does not measure how long it takes to process an event batch.

Solution: Add a new histogram metric event_batch_processing_milliseconds, 
that measures the time it takes to process an event batch. Also adds a debug 
log statement with the same information so we can debug spikes in processing time.