Only update a resource's status if it has changed (#1151)
Problem: NGF updates the status of resources even when the
status has not changed. When there are a large number of 
HTTPRoutes this can significantly decrease processing time.

Solution: Modify the status updater to only update status 
when it has changed.

Testing: Re-ran the scale tests. The scale test performance numbers 
significantly improved with this change.