Fix conformance release workflow (#1198)
Problem: The job used in the conformance tests to wait for the release to exist wasn't working properly. It only saw jobs in progress, and not jobs that are pending, which is what we needed to wait for.

Solution: Removed the action and replaced with a basic polling loop to ensure the NGF images exist before continuing with the tests.