Add stubbed NginxProxy CRD and watcher (#1216)
Problem: We have two upcoming stories that require the new NginxProxy CRD in order to be worked on.

Solution: To prevent one story from being blocked by the other, this work adds the new CRD with a blank spec, a controller that watches for it and triggers an update, and sets the status on the resource.