Update Gateway Architecture document to fit standard NGINX Formatting
This commit updates the architecture document to be more specific in its
titling, to add the F5 prefix to the description, and to change the
formatting used in the document to match our markdown conventions.

There are some other, smaller changes included to rephrase or change the
text in order to match the newer phrasing and architecture of the
equivalent NGINX Ingress Controller document for consistency.