Fix Helm tests on release tag (#1255)
Fix Helm tests on release tag and replace cut commands.

Problem: Helm tests fail on release tag because the kind load command did not work when the provided images spanned multiple lines. Also, cut commands did not work as input was not a single value.

Solution: Instead of using steps.*.outputs.tags which can output multiple lines, changed to use the hardcoded path to the image and a specific version.