Add basic system test with utilities (#1274)
Problem: In order to test a full NGF system running in k8s, we need a framework that can easily deploy apps and send traffic.

Solution: Enhance the framework with functions to create apps and send traffic using port forwarding. Also added a basic test to utilize these functions as a proof of concept.