Set SupportedVersion on GatewayClass (#1301)
Problem: NGF does not set the SupportedVersion condition 
on the GatewayClass.

Solution: Set the SupportedVersion condition on the GatewayClass.

This PR adds a new controller that watches the metadata of CRDs.
CRD events are filtered using a custom predicate that inspects the 
gateway.networking.k8s.io/bundle-version annotation. CRDs without 
this annotation will be ignored. Updates to this annotation will trigger 
a state change and build a new graph. This required some changes to 
how we determine if the state has changed in the changeTrackingUpdater.