Use latest stable version of Go
Problem: Since Go changed the versioning in go.mod to specify a patch
release in 1.21, we don't use the latest go version when building the
binary anymore.
Before with 1.20 in go.mod we would use the latest version of 1.20.x,
but now since we specify 1.21.3 in go.mod we use that fixed version
and not 1.21.4 for example, and future updates.

Solution: Update the pipeline to always use the latest stable version of
Go in the pipeline.