Prototype ControlPlane DataPlane split using a custom agent
This commit splits the current Deployment of pods with collocated
Control plane and Data plane into two Deployments:
- Control plane Deployment
- Data plane Deployment

A Control Plane Pod includes a single container that runs the NGF
static mode, which exposes a gRPC API for Data planes to
stream configuration changes.

A Data Plane Pod includes two containers:
- NGINX container (same as before in the previous collocated CP/DP pod)
- Agent container, which includes an agent that gets configuration from
the Control Plane via gRPC and applies it to NGINX.

The agent is the same code as was previously in the collocated CP/DP pod
with an addition of streaming of configuration via gRPC.
The agent also includes NGINX/NGINX Plus Prometheus metrics, which were
moved from the static mode.

The agent is in the same binary as the static mode, and the same
container image is used.  The agent can be run via the new command
"agent".

By default, the agent can only receive payloads of max 4MB. However,
this can be increased via https://pkg.go.dev/google.golang.org/grpc#MaxRecvMsgSizeCallOption
Experimental option.

gRPC communication layer consists of:
- server (control plane) library -  internal/grpc/sdk/server
- client (agent) library - internal/grpc/sdk/client/client
- protobuf definitions - internal/grpc/controlplane/controlplane.pb.go
- server and client cmd apps for quick local experimentation that use
the libraries internal/grpc/cmd

The agent uses the mentioned client library.
The static mode uses the server library.

This is a PoC, so a lot of important things are out of scope:
- Authentication
- Horizontal scaling of Control plane
- Support for N+ API when re-configuring Data planes
- Sending configuration diffs rather than whole files
- Updating the provisioner mode.

Testing:
- HTTPS example -- https://github.com/nginxinc/nginx-gateway-fabric/tree/5e0d72357f5e6240fcbf311e00e6aa0067d67ce0/examples/https-termination
- Scalability test for HTTPS listeners -- https://github.com/nginxinc/nginx-gateway-fabric/blob/5e0d72357f5e6240fcbf311e00e6aa0067d67ce0/tests/scale/scale.md#scale-https-listeners-to-max-of-64

Some insights:
- Changes to the handler of the static mode are minimal. This was made
possible by the simplistic Control plane library.
- It is not difficult to transform the current Agent (which is part of
the static mode and which transfers files via shared volumes rather than
gRPC) to enable gRPC streaming of the configuration files changes.