Move generation check from change_processor to controller
Problem: When processing updates to cluster resources, for some
resources, we check their generation, so that we don't trigger state
change (graph rebuild) if the generation didn't change. This is a
performance optimization so that we don't rebuild the graph and as a
result do not regenerate NGINX config and reload it. This is not a
K8s-native solution, and may introduce complexity for the codebase.

Solution: Use `GenerationChangedPredicate` in controller-runtime to
filter out the resource events at the controller level.