Add ephemeral debugging guide (#1202)
Problem: No documentation on how to debug NGF in Kubernetes

Solution: Add a developer guide with step-by-step instructions 
on how to inject an ephemeral dlv debug container into NGF while
it's running in a cluster. Includes makefile targets to make the 
process easier and a dockerfile for the dlv debug image.