Add Helm Chart configuration table reference to installation guide
This commit adds a prominent section to the documentation's Helm
installation guide to direct a reader to the Helm README which has all
of its configurable parameters.

This information is being referenced instead of duplicated to avoid
drift, which has the side-effect of making the Helm chart itself a
single source of truth.