Test automation: zero-downtime upgrades (#1438)
Problem: Running system tests manually is a time consuming and arduous process, and can lead to inconsistent results.

Solution: Add automation for the zero-downtime upgrades test. This test must be run on GKE. The test deploys the previous released version of NGF, then begins sending http and https traffic while performing an upgrade to NGF. Results are rendered and written.

CSV files are getting too large to commit to the repository, so not including those going forward.