Update guidance and tooling for documentation (#1431)
These changes add a README file to the site subfolder instructing a
reader on how to build and contribute to the documentation. It removes
a similar README from the developer documentation folder while
incorporating a number of improvements.

With this commit there are also changes to the Makefile, with the aim
of clarifying the intent and usage of certain commands. Finally, the
new README references and is referenced by the general CONTRIBUTION
guidelines to reflect the standardised processes and reinforce our
"Docs like Code" approach.

Co-authored-by: Saylor Berman <s.berman@f5.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>