EndpointSlice resources tracked by Graph (#1432)
Problem: EndpointSlice resources are currently being tracked by the relationship capturer,
but to reduce the number of components we would like to track the resources in the Graph.

Solution: Enable Graph to track EndpointSlice resources and Services. In addition, this PR
refactors backend_refs to store the NamespacedName of a Service and a ServicePort instead
of the actual Service. This is done to track non-existing referenced services. Also, refactors
and removes the alwaysTruePredicate. Removes RelationshipCapturer.