Fix markdownlint in `site` (#1499)
Problem: The configuration for the `site` directory didn't have a
standard name and wasn't getting picked up by markdownlint-cli. The
Makefile was also referencing .markdownlint.yaml that doesn't exist.

Also the configuration for .pre-commit-config.yaml was invalid and the
autoupdate wasn't working correctly.

Solution: Change the name of the configuration to a standard one that
can be used automatically be the tool.
Update pre-commit config and Makefile and run the linter to fix the
existing problems.