Run unit tests with ginkgo (#1517)
Problem: We run our unit tests with go test, but many 
of our tests are written with ginkgo. While the ginkgo 
tests still run with go test, we are missing out on many 
of ginkgo features -- such as randomized test execution.

Solution: Run unit tests using the ginkgo CLI instead of go test. 
Due to a bug in the Cobra library, the tests in the cmd package 
cannot be run with ginkgo and are still run with go test. This 
results in two separate cover profiles: cmd-cover.out and cover.out.