Update package for CVE-2023-52425 (#1569)
Problem: There's a new CVE in the base image, but since Docker is
caching the layer, `apk update && apk upgrade` doesn't have any effect.

Solution: Only upgrade the affected package. This will invalidate the
cache and only upgrade the package with a CVE. This also makes the build
somewhat reproducible since we're not updating all the packages at build
time.