Add release-note section in PR template
Problem: Writing release notes at release time takes a long time and
it's not easy to do if you don't have as much context as the submitter
or simply too much time has passed.

Solution: Add a release notes section to be filled out when opening a
PR. This way at release time we can just grab all the descriptions and
speed up the process.