Remove "build images" link from README (#1575)
The Getting Started guide refers to deploying and installing NGF, which doesn't require building an image. This link doesn't make sense to have after these steps, and that guide is mainly written  as an alternate option for devs. In the context of Getting Started, it feels unnecessary, and can be found through the docs naturally if needed.