Add flag to opt out of product telemetry
Problem: As a user of NGF
I want an easy option to opt out of recording and sending of all telemetry data
So that if I am not comfortable sending that information, I can still continue to use NGF

Solution: Added a new configurable flag that can disable telemetry. On by default. Also turns off any RBAC if not needed (for N+ usage or telemetry)