Add support for ResponseHeaderModifier for HTTPRouteRule objects
Problem: Users want to add, set, and remove response headers.

Solution: Use add_header and proxy_hide_header NGINX directives to support ResponseHeaderModifier.

If the action is set, we configure the proxy_hide_header directive to remove the header if exists and add_header directive with the given value in the HTTPRoute spec.
If the action is remove, we configure the proxy_hide_header directive to remove the header.
If the action is add, we simply configure the add_header directive with the given value in the HTTPRoute spec.