Add Codecov (#1626)
Problem: We don't have a history of our code coverage and we want to see
how it changes over time and with every PR. We also want to have a badge
that shows the overall coverage.

Solution: Use Codecov to upload coverage and get diffs in every PR that
changes code. Also add a badge in the README to show the coverage.