Force conformance tests to fail properly
Problem: If the conformance test Pod runner failed, the `make` process would still pass. This is because the failure code would not propagate to outside of the Pod.

Solution: After the Pod runs, check the exit code and fail it it's non-zero. Used a bash script for this because running the `kubectl get pod` command in the Makefile would not return an Errored Pod for some reason. Only worked in bash.