POC - Refactor Status Updater to be resource agnostic
Problem:
Currently, the status updater is aware of what resource it is updating
the status of. This makes it difficult to extend because for each new
resource we add support for, we have to modify the status updater.

Solution:
Replace old status updater with two new ones

(1) Regular Updater, to update statuses of multiple resources via
UpdateRequest type

(2) CachingGroupUpdater to update statuses of groups of resources and
cache results until the updater is enabled (when pod becomes leader).
It uses Regular Updater.

Using groups allow replacing statuses of subset of resources, without
the need to recompute all cached statuses.

The new status2 package (which will replace status package) is
resource agnostic. This is acomplished by representing status update
as a function (Setter).

The manager packages were updated:
- provisioner mode, to use regular Updater
- static mode, to use CachingGroupUpdater

status Setters were updated and moved to the static package.

CLOSES -- https://github.com/nginxinc/nginx-gateway-fabric/issues/1071