Add ClientSettingsPolicy CRD (#1793)
Problem: Users need a way to configure the behavior 
between the downstream client and NGF. For example, 
be able to set the client max body size.

Solution: Add the ClientSettingsPolicy CRD, which 
provides a way to configure this behavior. 
Note: this PR contains the CRD only. 
A subsequent PR will add the implementation.