Add NginxProxy CRD (#1815)
Problem: Users want to be able to configure global Gateway  settings, such as the Otel tracing exporter, for all Gateways in a Class.

Solution: Add the NginxProxy CRD, which provides a way to configure these settings.
Note: this PR contains the CRD only. A subsequent PR will add the implementation.