Add ObservabilityPolicy CRD (#1848)
Problem: Users want to be able to configure observability settings for their HTTPRoutes, such as tracing.

Solution: Add the ObservabilityPolicy CRD, which is a direct policy that will attach to HTTPRoutes to configure these settings. Note: this pR contains the CRD only. A subsequent PR will add the implementation.

Also removed some regex restrictions in the NginxProxy CRD.